pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-financial-market-env"
version = "0.1.0"
description = "Financial Market trading environment for OpenEnv — NIFTY 50 stocks with 3 difficulty tiers"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.1",
    "yfinance>=0.2.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.25.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["server"]
